# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MemeHub API - 9GAG Clone
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts:
    get:
      operationId: getPosts
      tags: [posts]
      summary: Get posts feed
      parameters:
        - name: section
          in: query
          schema:
            type: string
            enum: [hot, trending, fresh, top]
            default: hot
        - name: tag
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: number
            default: 1
        - name: limit
          in: query
          schema:
            type: number
            default: 10
      responses:
        "200":
          description: Posts list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostsListResponse"

    post:
      operationId: createPost
      tags: [posts]
      summary: Create a new post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostRequest"
      responses:
        "201":
          description: Post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/users:
    get:
      operationId: getAdminUsers
      tags: [admin]
      summary: Get all users (admin)
      parameters:
        - name: page
          in: query
          schema:
            type: number
            default: 1
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUsersListResponse"
  /admin/users/{id}/ban:
    post:
      operationId: banUser
      tags: [admin]
      summary: Ban or unban user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BanRequest"
      responses:
        "200":
          description: User banned/unbanned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/users/{id}/role:
    put:
      operationId: updateUserRole
      tags: [admin]
      summary: Update user role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRoleRequest"
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/posts:
    get:
      operationId: getAdminPosts
      tags: [admin]
      summary: Get all posts (admin)
      parameters:
        - name: page
          in: query
          schema:
            type: number
            default: 1
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [all, pending, approved, removed]
            default: all
      responses:
        "200":
          description: Posts list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPostsListResponse"
  /admin/posts/{id}/status:
    put:
      operationId: updatePostStatus
      tags: [admin]
      summary: Update post status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePostStatusRequest"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/settings:
    get:
      operationId: getAdminSettings
      tags: [admin]
      summary: Get site settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"
    put:
      operationId: updateAdminSettings
      tags: [admin]
      summary: Update site settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteSettings"
      responses:
        "200":
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"
  /admin/tags:
    post:
      operationId: createTag
      tags: [admin]
      summary: Create a new tag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTagRequest"
      responses:
        "201":
          description: Tag created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
  /admin/tags/{id}:
    delete:
      operationId: deleteTag
      tags: [admin]
      summary: Delete a tag
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/comments:
    get:
      operationId: getAdminComments
      tags: [admin]
      summary: Get all comments (admin)
      parameters:
        - name: page
          in: query
          schema:
            type: number
            default: 1
      responses:
        "200":
          description: Comments list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentsListResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    Post:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        imageUrl:
          type: string
        gifUrl:
          type: string
          nullable: true
        type:
          type: string
          enum: [image, gif, video]
        upvotes:
          type: number
        downvotes:
          type: number
        points:
          type: number
        commentsCount:
          type: number
        viewsCount:
          type: number
        status:
          type: string
          enum: [pending, approved, removed]
        userVote:
          type: string
          enum: [up, down]
          nullable: true
        isSaved:
          type: boolean
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
        author:
          $ref: "#/components/schemas/PostAuthor"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - imageUrl
        - type
        - upvotes
        - downvotes
        - points
        - commentsCount
        - viewsCount
        - status
        - isSaved
        - tags
        - author
        - createdAt
    PostAuthor:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        avatar:
          type: string
          nullable: true
      required:
        - id
        - username
    PostsListResponse:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        total:
          type: number
        page:
          type: number
        totalPages:
          type: number
        hasMore:
          type: boolean
      required:
        - posts
        - total
        - page
        - totalPages
        - hasMore
    Comment:
      type: object
      properties:
        id:
          type: string
        content:
          type: string
        upvotes:
          type: number
        downvotes:
          type: number
        userVote:
          type: string
          enum: [up, down]
          nullable: true
        author:
          $ref: "#/components/schemas/PostAuthor"
        parentId:
          type: string
          nullable: true
        replies:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - content
        - upvotes
        - downvotes
        - author
        - createdAt
    CommentsListResponse:
      type: object
      properties:
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        total:
          type: number
        page:
          type: number
        totalPages:
          type: number
      required:
        - comments
        - total
        - page
        - totalPages
...
    Tag:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        postsCount:
          type: number
        color:
          type: string
          nullable: true
      required:
        - id
        - name
        - slug
        - postsCount
...
    CreateTagRequest:
      type: object
      properties:
        name:
          type: string
        color:
          type: string
          nullable: true
      required:
        - name
    CreatePostRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 200
        imageUrl:
          type: string
        type:
          type: string
          enum: [image, gif, video]
        tagIds:
          type: array
          items:
            type: string
      required:
        - title
        - imageUrl
        - type
...
    AdminUser:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        avatar:
          type: string
          nullable: true
        role:
          type: string
          enum: [user, moderator, admin]
        isBanned:
          type: boolean
        postsCount:
          type: number
        commentsCount:
          type: number
        totalPoints:
          type: number
        createdAt:
          type: string
          format: date-time
        lastActiveAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - username
        - email
        - role
        - isBanned
        - postsCount
        - commentsCount
        - totalPoints
        - createdAt
    AdminUsersListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
        total:
          type: number
        page:
          type: number
        totalPages:
          type: number
      required:
        - users
        - total
        - page
        - totalPages
    AdminPostsListResponse:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        total:
          type: number
        page:
          type: number
        totalPages:
          type: number
      required:
        - posts
        - total
        - page
        - totalPages
    BanRequest:
      type: object
      properties:
        reason:
          type: string
          nullable: true
        isBanned:
          type: boolean
      required:
        - isBanned
    UpdateRoleRequest:
      type: object
      properties:
        role:
          type: string
          enum: [user, moderator, admin]
      required:
        - role
    UpdatePostStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [pending, approved, removed]
      required:
        - status
    SiteSettings:
      type: object
      properties:
        siteName:
          type: string
        siteDescription:
          type: string
        allowRegistration:
          type: boolean
        requireApproval:
          type: boolean
        maxUploadSizeMb:
          type: number
        allowedFileTypes:
          type: array
          items:
            type: string
        huggingFaceRepo:
          type: string
        maintenanceMode:
          type: boolean
      required:
        - siteName
        - siteDescription
        - allowRegistration
        - requireApproval
        - maxUploadSizeMb
        - allowedFileTypes
        - huggingFaceRepo
        - maintenanceMode