# lib/api-zod/src/generated/types/siteSettings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MemeHub API - 9GAG Clone
 * OpenAPI spec version: 0.1.0
 */

export interface SiteSettings {
  siteName: string;
  siteDescription: string;
  allowRegistration: boolean;
  requireApproval: boolean;
  maxUploadSizeMb: number;
  allowedFileTypes: string[];
  huggingFaceRepo: string;
  maintenanceMode: boolean;
}